Tools project maintenance. Compiler warning fixed.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8037 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/Session.cpp
@@ -2845,7 +2845,7 @@ void idSessionLocal::Frame() {
 			//stgatilov #4865: impose artificial FPS limit
 			static int64_t prevUsec = INT64_MIN;
 			int64_t neededDelta = 1000000 / com_maxFPS.GetInteger();
-			while (Sys_GetTimeMicroseconds() - prevUsec < neededDelta) {
+			while ((int64_t)Sys_GetTimeMicroseconds() - prevUsec < neededDelta) {
 				//note: this is busy-wait loop
 				#ifdef __SSE2__
 				_mm_pause();
